Add schema support for OpenSearchAdapter
Also preps for query parsing with OpenSearch
