+++ benchmarks/src/main/java/com/slack/kaldb/QueryBenchmark.java
@@ -109,7 +109,7 @@ public class QueryBenchmark {
     logStore.commit();
     logStore.refresh();
     logIndexSearcher =
-        new LogIndexSearcherImpl(logStore.getSearcherManager(), logStore.getSchema());
+        new LogIndexSearcherImpl(logStore.getSearcherManager(), logStore.getSchema(), false);
   }
 
   @TearDown(Level.Trial)
