+++ kaldb/src/main/java/com/slack/kaldb/elasticsearchApi/ElasticsearchApiService.java
@@ -23,7 +23,7 @@ import com.slack.kaldb.elasticsearchApi.searchResponse.HitsMetadata;
 import com.slack.kaldb.elasticsearchApi.searchResponse.SearchResponseHit;
 import com.slack.kaldb.elasticsearchApi.searchResponse.SearchResponseMetadata;
 import com.slack.kaldb.logstore.LogMessage;
-import com.slack.kaldb.logstore.opensearch.OpenSearchAggregationAdapter;
+import com.slack.kaldb.logstore.opensearch.OpenSearchInternalAggregation;
 import com.slack.kaldb.proto.service.KaldbSearch;
 import com.slack.kaldb.server.KaldbQueryServiceBase;
 import com.slack.kaldb.util.JsonUtil;
@@ -145,7 +145,7 @@ public class ElasticsearchApiService {
 
   private JsonNode parseAggregations(ByteString byteInput) throws IOException {
     InternalAggregation internalAggregations =
-        OpenSearchAggregationAdapter.fromByteArray(byteInput.toByteArray());
+        OpenSearchInternalAggregation.fromByteArray(byteInput.toByteArray());
     if (internalAggregations != null) {
       return objectMapper.readTree(internalAggregations.toString());
     }
