+++ kaldb/src/test/java/com/slack/kaldb/logstore/LuceneIndexStoreImplTest.java
@@ -408,7 +408,8 @@ public class LuceneIndexStoreImplTest {
       LogIndexSearcherImpl newSearcher =
           new LogIndexSearcherImpl(
               LogIndexSearcherImpl.searcherManagerFromPath(tempFolder.getRoot().toPath()),
-              logStore.getSchema());
+              logStore.getSchema(),
+              false);
       Collection<LogMessage> newResults =
           findAllMessages(newSearcher, MessageUtil.TEST_DATASET_NAME, "Message1", 100);
       assertThat(newResults.size()).isEqualTo(1);
@@ -451,7 +452,8 @@ public class LuceneIndexStoreImplTest {
       LogIndexSearcherImpl newSearcher =
           new LogIndexSearcherImpl(
               LogIndexSearcherImpl.searcherManagerFromPath(tempFolder.getRoot().toPath()),
-              logStore.getSchema());
+              logStore.getSchema(),
+              false);
 
       Collection<LogMessage> newResults =
           findAllMessages(newSearcher, MessageUtil.TEST_DATASET_NAME, "Message1", 100);
